requirements.txt
python-dotenv==1.1.1
aiogram==3.22.0
SQLAlchemy==2.0.43
alembic==1.16.5